+++ UriTemplate/Template.php
@@ -23,8 +23,8 @@ use function array_reduce;
 use function array_unique;
 use function preg_match_all;
 use function preg_replace;
-use function str_contains;
 use function str_replace;
+use function strpbrk;
 
 use const PREG_SET_ORDER;
 
@@ -65,9 +65,7 @@ final class Template implements Stringable
         $template = (string) $template;
         /** @var string $remainder */
         $remainder = preg_replace(self::REGEXP_EXPRESSION_DETECTOR, '', $template);
-        if (str_contains($remainder, '{') || str_contains($remainder, '}')) {
-            throw new SyntaxError('The template "'.$template.'" contains invalid expressions.');
-        }
+        false === strpbrk($remainder, '{}') || throw new SyntaxError('The template "'.$template.'" contains invalid expressions.');
 
         preg_match_all(self::REGEXP_EXPRESSION_DETECTOR, $template, $founds, PREG_SET_ORDER);